requirements.txt
Flask==3.0.3
flask-cors==4.0.1
scikit-learn==1.4.0
textblob==0.17.1
nltk==3.8.1
pandas==2.2.0
numpy==1.26.4
keras==2.15.0
Pillow==10.2.0
h5py==3.10.0
tensorflow==2.15.0
